Simplified the build process
Greatly simplified the build process by removing the requirement to
configure dependencies externally. The desktop app can now be built
with qmake and make alone, since all dependencies are included directly
in the Silverlock repository. There is now a single project-level
maintenance script in the bin directory which is used to perform
various maintenance operations, such as updating dependencies to their
newest versions, updating translation files, and whatever is necessary
to include in the future.

Also, the amalgamation version of Botan is now compiled directly into
libsilverlock, which might increase efficiency a little.